Create function to randomly choose 3 words (Rock, Paper or Scissor)
* Define an array that contains the words.Write a function that generates a random number
within the range of the array's length. This random number will be used as the index to select
a random word from the array.

* Inside the function, use the Math.random() function to generate a random decimal number between 0 and 1.
Multiply this number by the length of the array to get a random index value.

Use the Math.floor() function to round down the decimal number to the nearest integer, which will give
you a whole number within the valid index range of the array.

Return the word at the randomly generated index from the array.

To see the random word, assign the return value of the function to a variabled then print it  to the console.
This allows you to  store and display it.